+++ tests/compilers/issues.rs
@@ -436,6 +436,8 @@ fn large_number_local(mut config: crate::Config) -> Result<()> {
           i64.add
           local.get 16
           i64.add
+          local.get 512
+          i64.add
         )
       )
     "#;
@@ -452,6 +454,8 @@ fn large_number_local(mut config: crate::Config) -> Result<()> {
     Ok(())
 }
 
+// TODO: the tests fails on RISC-V as the `j` instruction can reach only +- 1MiB offset.
+#[cfg(any(target_arch = "x86_64", target_arch = "aarch64"))]
 #[compiler_test(issues)]
 /// Singlepass panics on aarch64 for long relocations.
 ///
@@ -468,6 +472,8 @@ fn issue_4519(mut config: crate::Config) -> Result<()> {
     Ok(())
 }
 
+// TODO: the tests fails on RISC-V as the `j` instruction can reach only +- 1MiB offset.
+#[cfg(any(target_arch = "x86_64", target_arch = "aarch64"))]
 #[compiler_test(issues)]
 /// Singlepass panics on aarch64 for long relocations.
 /// This test specifically targets the emission of sdiv64, srem64, urem64 binops.
@@ -686,6 +692,56 @@ fn issue_6004_exception(mut config: crate::Config) {
     let result = instance
         .exports
         .get_function("throw-expect-42")
+        .unwrap()
+        .call(&mut store, &[])
+        .unwrap();
+    assert_eq!(&Value::I32(42), result.first().unwrap());
+}
+
+#[cfg(not(target_os = "windows"))]
+#[compiler_test(issues)]
+fn issue_5719_shared_catch_clause_block(mut config: crate::Config) {
+    let wasm_bytes = wat2wasm(
+        r#"
+(module
+    (tag $err (param))
+    (tag $err2 (param))
+    (export "f" (func $f))
+    (func $f (result i32)
+        block
+            block
+                try_table (catch $err 0) (catch $err2 0)
+                    throw $err2
+                end
+                unreachable
+            end
+            i32.const 42
+            return
+        end
+        i32.const 13
+        return
+    )
+)
+"#
+        .as_bytes(),
+    )
+    .expect("wat2wasm must succeed");
+
+    let mut store = config.store();
+    let module = match Module::new(&store, wasm_bytes) {
+        Err(CompileError::Validate(message))
+            if message.contains("exceptions proposal not enabled") =>
+        {
+            // Skip the test in that case.
+            return;
+        }
+        Ok(module) => module,
+        _ => unreachable!(),
+    };
+    let instance = Instance::new(&mut store, &module, &imports! {}).unwrap();
+    let result = instance
+        .exports
+        .get_function("f")
         .unwrap()
         .call(&mut store, &[])
         .unwrap();